Update on Frontend_zf - Added live client update when a client's room is judged - Moved judging into its own module in ./server/judging.js
